web_app/requirements.txt
altair
pandas
psycopg2-binary==2.9.3
streamlit
sqlalchemy